+++ src/com/dparker/apps/checkvalve/UserVisibleMessage.java
@@ -19,7 +19,9 @@
 
 package com.dparker.apps.checkvalve;
 
+import java.io.File;
 import android.content.Context;
+import android.content.Intent;
 import android.util.Log;
 import android.widget.Toast;
 
@@ -36,7 +38,7 @@ public class UserVisibleMessage {
      * @param text The text to be displayed in the message
      */
     public static Toast showMessage( Context context, CharSequence text ) {
-        Log.d(TAG, "Showing message [context=" + context.toString() + "][text=" + text + "]");
+        Log.d(TAG, "showMessage(): Showing message [context=" + context.toString() + "][text=" + text + "]");
 
         // Create a new Toast message
         Toast t = Toast.makeText(context, text, Toast.LENGTH_SHORT);
@@ -51,11 +53,37 @@ public class UserVisibleMessage {
      * @param resId The ID of the string resource to be displayed in the message
      */
     public static Toast showMessage( Context context, int resId ) {
-        Log.d(TAG, "Showing message [context=" + context.toString() + "][resId=" + resId + "]");
+        Log.d(TAG, "showMessage(): Showing message [context=" + context.toString() + "][resId=" + resId + "]");
 
         // Create a new Toast message
         Toast t = Toast.makeText(context, resId, Toast.LENGTH_SHORT);
         t.show();
         return t;
+    }
+    
+    /**
+     * Displays a note to the user and includes the option to not show this note again.
+     * 
+     * @param context The context to use
+     * @param file The filename of the marker file to create or delete
+     * @param resId The resource ID of the message to display in the note
+     */
+    public static void showNote( Context context, String file, int resId ) {
+        File f = new File(context.getFilesDir(), file);
+        
+        if( f.exists() ) {
+            Log.d(TAG, "showNote(): Marker file " + file + " exists; note will not be displayed.");
+            return;
+        }
+        
+        Log.d(TAG, "showNote(): Marker file " + file + " does not exist; showing note.");
+        
+        
+        Intent showNoteIntent = new Intent();
+        showNoteIntent.setClassName("com.dparker.apps.checkvalve", "com.dparker.apps.checkvalve.ShowNoteActivity");
+        showNoteIntent.setFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT);
+        showNoteIntent.putExtra(Values.EXTRA_NOTE_ID, resId);
+        showNoteIntent.putExtra(Values.EXTRA_FILE_NAME, file);
+        context.startActivity(showNoteIntent);
     }
 }